update: just a minute change, meta title, keywords and desc is added into the single press page, which i missed earlier to add.
so now, since already we have minimised the css code, and then i had done this change of metatitle and all, so now uptill

this commit, i have made the press page ready and also minimised the css and also backup of the prev css files have been taken and stored locally which goes by the file name mentioning the dt: 11th oct. (so directly deplyoing this).

also added the direction in the footer for press releases

now i have also resolved dangerousInnerHtml tag to render the press.content

also incresed the word limit to show in the intro part from 150 to 550 words